Start changes to otherPagesCSS
Removed contactSpacer and made it a general spacer for all the pages (better name).

Replaced all the HTML that used the old contactSpacer class to spacer.

Replaced HTML class of indexText to a new name pageTitle (better name) and started the process to make that fluid typography.

Added back animation fly in on the portfolio page. Not sure why it was removed.

Must work on cleaning the rest of the CSS, add new and better names, and implement perfect fluid typography.