weird issue fix
Attempt to remove unneeded comments in main.css caused the about and contact modals to disappear.

Traced the issue back to what seems to be the w3-modal class. Commented it out and everything seems to be back to normal.

Issue was fixed but not entirely sure why that caused it.